+++ TeamProject_HONJAL/Honjal/src/main/java/com/honjal/honjal/model/ContentVO.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @Builder
 public class ContentVO {
 
-	private Long content_num;		// 글번호
+	private Integer content_num;		// 글번호
 	private Integer member_num;		// 작성자 번호
 	private String member_nname;	// 작성자 닉네임
 	private String board_code;		// 게시판코드
